chat: refactor to host-provided renderEntityCard pattern
- Delete object-card.tsx — host now owns entity rendering entirely
- Add chat-ref.types.ts (wire-schema only)
- chat-message-enhanced: accept renderEntityCard?: (ref) => ReactNode
- Update memo equality + drop canDiscussType/onCardDiscuss in favor of single renderer prop
- Export HoverCard from components/index.ts for hub composition

Per multi-platform-hub user feedback: OSS-lib stays data-agnostic, host
(multi-platform-hub) owns the type→icon mapping, slash-command gating,
and visual chrome. Single source of truth lives in the host.
