+++ openframe-frontend-core/src/components/chat/chat-message-list.tsx
@@ -9,7 +9,7 @@ import type { ChatMessageListProps } from "./types"
 const BOTTOM_THRESHOLD = 30 // px from bottom to be considered "at bottom"
 
 const ChatMessageList = forwardRef<HTMLDivElement, ChatMessageListProps>(
-  ({ className, messages, dialogId, isLoading = false, isTyping = false, autoScroll = true, showAvatars = true, contentClassName, assistantType, assistantIcon, pendingApprovals, hasNextPage, isFetchingNextPage, onLoadMore, canDiscussType, onCardDiscuss, ...props }, ref) => {
+  ({ className, messages, dialogId, isLoading = false, isTyping = false, autoScroll = true, showAvatars = true, contentClassName, assistantType, assistantIcon, pendingApprovals, hasNextPage, isFetchingNextPage, onLoadMore, renderEntityCard, ...props }, ref) => {
     const scrollRef = useRef<HTMLDivElement>(null)
     const sentinelRef = useRef<HTMLDivElement>(null)
     const onLoadMoreRef = useRef(onLoadMore)
@@ -309,8 +309,7 @@ const ChatMessageList = forwardRef<HTMLDivElement, ChatMessageListProps>(
                 authorType={message.authorType}
                 assistantIcon={message.role !== 'user' ? assistantIcon : undefined}
                 chatRefs={message.chatRefs}
-                canDiscussType={canDiscussType}
-                onCardDiscuss={onCardDiscuss}
+                renderEntityCard={renderEntityCard}
               />
             ))}
           </div>
