+++ openframe-frontend-core/src/components/chat/types/component.types.ts
@@ -5,7 +5,7 @@
 import type { ButtonHTMLAttributes, ComponentType, HTMLAttributes, TextareaHTMLAttributes } from 'react'
 import type { AssistantType, AuthorType, ChatApprovalStatus, ConnectionStatus } from './chat.types'
 import type { ApprovalRequestData, Message, MessageSegment, ToolExecutionData } from './message.types'
-import type { ChatRef } from '../object-card'
+import type { ChatRef } from '../chat-ref.types'
 
 // ========== Chat Container Props ==========
 
@@ -59,26 +59,25 @@ export interface ChatMessageEnhancedProps extends Omit<HTMLAttributes<HTMLDivEle
   onApprove?: (requestId?: string) => void | Promise<void>
   onReject?: (requestId?: string) => void | Promise<void>
   /**
-   * Per-row metadata for inline object-card rendering (v6.1 §B.2.6+§B.2.7).
-   * Keyed by `<documentType>:<primaryKey>`. When present, text segments are
-   * passed through the `remarkCardLinks` plugin and `[card://<type>:<id>]`
-   * markers expand into <ObjectCard /> hover pills. When undefined or empty,
-   * messages render as plain markdown (backward compat).
+   * Per-row metadata for inline entity-card rendering (v6.1 §B.2.6+§B.2.7).
+   * Keyed by `<documentType>:<primaryKey>`. When present AND
+   * `renderEntityCard` is also provided, text segments are passed through
+   * the `remarkCardLinks` plugin and `[card://<type>:<id>]` markers
+   * expand into the host's chosen inline component. When unset (or
+   * `renderEntityCard` unset), messages render as plain markdown.
    */
   chatRefs?: Record<string, ChatRef>
   /**
-   * Whether a given card type supports the "Discuss" action. Defers to the
-   * host's slash-command registry per v6.1 DRY duplications #2 — the OSS-lib
-   * has no knowledge of registered commands.
+   * Host-provided renderer for inline entity cards (v6.1 §B.2.7 — DRY
+   * duplications #2). The OSS-lib delegates all entity-specific rendering
+   * (type→icon mapping, hover-card chrome, action buttons, slash-command
+   * gating) to the host so the library stays free of app-specific
+   * knowledge.
+   *
+   * Return `null` for any ref the host can't render — the renderer falls
+   * back to plain text title-only.
    */
-  canDiscussType?: (type: string) => boolean
-  /**
-   * Click handler for the "Discuss" button on an ObjectCard. The host
-   * synthesizes a natural-prompt turn ("Tell me more about <title>") with a
-   * structured `commandOverride.entityIdFilter` field on the request body.
-   * v6.1 §B.2.8.
-   */
-  onCardDiscuss?: (reference: ChatRef) => void
+  renderEntityCard?: (reference: ChatRef) => React.ReactNode
 }
 
 // ========== Chat Message List Props ==========
@@ -102,15 +101,9 @@ export interface ChatMessageListProps extends HTMLAttributes<HTMLDivElement> {
   hasNextPage?: boolean
   isFetchingNextPage?: boolean
   onLoadMore?: () => void
-  /** Whether a given object-card type supports the "Discuss" action. The
-   *  host queries its slash-command registry (or equivalent) per v6.1
-   *  DRY duplications #2 — the OSS-lib has no knowledge of registered
-   *  commands. Forwarded to every message's ChatMessageEnhanced. */
-  canDiscussType?: (type: string) => boolean
-  /** Click handler for the Discuss button on an ObjectCard. The host
-   *  synthesizes a natural-prompt turn ("Tell me more about <title>")
-   *  with a structured `commandOverride.entityIdFilter`. v6.1 §B.2.8. */
-  onCardDiscuss?: (reference: ChatRef) => void
+  /** Host-provided renderer for inline entity cards. Forwarded verbatim
+   *  to every message's ChatMessageEnhanced. v6.1 §B.2.7. */
+  renderEntityCard?: (reference: ChatRef) => React.ReactNode
 }
 
 export interface ChatMessageListRef {
